requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
ultralytics>=8.0.0
pytesseract>=0.3.8
pyttsx3>=2.90
SpeechRecognition>=3.8.1
vosk>=0.3.45
pyaudio>=0.2.11
pipwin>=0.5.2
pygame>=2.0.0
tensorflow>=2.5.0
tensorflow-hub>=0.12.0
aiohttp>=3.9.0
torch>=1.9.0
Pillow>=8.0.0
mediapipe>=0.9.0
sounddevice>=0.4.5
scipy>=1.7.0
python-rtmidi>=1.4.9
transformers>=4.30.0
easyocr>=1.7.0
# Core speech and language packages
pyttsx3>=2.90
SpeechRecognition>=3.8.1
langdetect>=1.0.9

# Optional offline speech recognition (VOSK)
vosk>=0.3.45
sounddevice>=0.4.5

# Vision and image processing
opencv-python>=4.5.0
numpy>=1.19.0

# Network requests and model downloads
requests>=2.26.0

# Optional audio/microphone backend
pyaudio>=0.2.11
pipwin>=0.5.2  # useful for installing pyaudio on Windows

# Optional: useful scientific libs (if needed for expansion)
scipy>=1.7.0
